backend/requirements.txt
Django>=3.2.0,<4.0.0
djangorestframework>=3.12.0
django-cors-headers>=3.10.0
django-filter>=21.1
python-dotenv>=0.19.0
pillow>=8.0.0
pytz>=2021.3
sqlparse>=0.4.2
psycopg2-binary>=2.9.0